Dang it IntelliJ!
Those casts matter!

+++ src/main/java/docs/GenerateFigures.java
@@ -931,8 +931,8 @@ public class GenerateFigures extends AppWindow {
                       ctrlHandle2.getY2() + cy2 * scale,
                       getDrawingPanel()
               );
-      components.put("interpolation" + 2 * i - 1, line1);
-      components.put("interpolation" + 2 * i, line2);
+      components.put("interpolation" + (int) (2 * i - 1), line1);
+      components.put("interpolation" + (int) (2 * i), line2);
       line1.setStrokeColor(Color.getHSBColor((float) scale, 1, 1));
       line2.setStrokeColor(line1.getStrokeColor());
     }
@@ -952,8 +952,8 @@ public class GenerateFigures extends AppWindow {
     for (int i = 1; i < interpolations; i++) {
       double scale = (double) i / interpolations;
       Line
-              line1 = (Line) components.get("interpolation" + 2 * i - 1),
-              line2 = (Line) components.get("interpolation" + 2 * i),
+              line1 = (Line) components.get("interpolation" + (int) (2 * i - 1)),
+              line2 = (Line) components.get("interpolation" + (int) (2 * i)),
               line3 = new Line(
                       line1.getX1() + (line1.getX2() - line1.getX1()) * scale,
                       line1.getY1() + (line1.getY2() - line1.getY1()) * scale,
@@ -961,7 +961,7 @@ public class GenerateFigures extends AppWindow {
                       line2.getY1() + (line2.getY2() - line2.getY1()) * scale,
                       getDrawingPanel()
               );
-      components.put("interpolation" + 2 * i - 1 + "-" + 2 * i, line3);
+      components.put("interpolation" + (int) (2 * i - 1) + "-" + 2 * i, line3);
 
       Ellipse ellipse = new Ellipse(
               line3.getX1() + (line3.getX2() - line3.getX1()) * scale - POINT_DIAMETER / 2,
